+++ portal_do_aluno/lib/features/admin/presentation/pages/gestao_escolar/diciplinas_page_page.dart
@@ -1,6 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:portal_do_aluno/features/admin/data/datasources/cadastrar_diciplina_firestore.dart';
+import 'package:portal_do_aluno/shared/helpers/show_confirmation_dialog.dart';
 import 'package:portal_do_aluno/shared/widgets/popup_menu_botton.dart';
 import 'package:portal_do_aluno/navigation/navigation_sevice.dart';
 import 'package:portal_do_aluno/navigation/route_names.dart';
@@ -80,10 +81,17 @@ class _DiciplinasPageState extends State<DiciplinasPage> {
                         MenuItemConfig(
                           value: 'Excluir',
                           label: 'Excluir',
-                          onSelected: (id, context, extra) {
+                          onSelected: (id, context, extra) async {
                             if (id != null) {
                               try {
-                                _disciplinaService.excluirDisciplina(id);
+                                final excluir = await showConfirmationDialog(
+                                  context: context,
+                                  title: 'Deseja Excluir essa Disciplina?',
+                                  content: 'Essa ação é irreversível',
+                                );
+                                if (excluir == true) {
+                                  _disciplinaService.excluirDisciplina(id);
+                                }
                               } catch (e) {
                                 if (e is Exception) {
                                   ScaffoldMessenger.of(context).showSnackBar(
